Refactoring C-style casts into C++-style 'static_cast'
... and other minor, related changes in code, such as removing
unnecessary casts.

+++ src/gui/file/basicfileview.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
- *   Copyright (C) 2004-2017 by Thomas Fischer <fischer@unix-ag.uni-kl.de> *
+ *   Copyright (C) 2004-2018 by Thomas Fischer <fischer@unix-ag.uni-kl.de> *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -230,7 +230,7 @@ void BasicFileView::headerActionToggled()
     QAction *action = qobject_cast<QAction *>(sender());
     if (action == nullptr) return;
     bool ok = false;
-    const int col = (int)action->data().toInt(&ok);
+    const int col = action->data().toInt(&ok);
     if (!ok) return;
 
     header()->setSectionHidden(col, !header()->isSectionHidden(col));
